classificationModel/requirements.txt
tensorflow-macos-2.9.0